chore: adjusted util structure
