+++ src/app/modules/signatory-progress/utils/intefaces.ts
@@ -1,0 +1,11 @@
+import { FacetsModel } from 'app/modules/signatory-progress/store/interface';
+
+export interface DataPoint {
+  x: string;
+  y: number | null;
+}
+
+export interface specPubsItemModel {
+  name: string;
+  specPub: Array<FacetsModel | null>;
+}
